requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2==2.8.6
stripe==2.63.0
drf-spectacular==0.26.2
djangorestframework-simplejwt==5.2.2
django-rest-swagger==2.2.0
celery[redis]==5.2.7
flower==1.2.0
python-dotenv==1.0.0
flake8==6.0.0

# requests==2.31.0
# futures==3.1.1
faker==18.10.1